+++ src/Fantomas.Core/SourceParser.fs
@@ -42,7 +42,7 @@ let rec (|RationalConst|) =
     function
     | SynRationalConst.Integer i -> string i
     | SynRationalConst.Rational (numerator, denominator, _) -> sprintf "(%i/%i)" numerator denominator
-    | SynRationalConst.Negate (RationalConst s) -> sprintf "- %s" s
+    | SynRationalConst.Negate (RationalConst s) -> sprintf "-%s" s
 
 let (|String|_|) e =
     match e with