+++ src/Fantomas.Core.Tests/SynConstTests.fs
@@ -141,7 +141,7 @@ namespace Krach
 
 module Runner =
 
-    let mPerSecond = 1000<m/second> // foo
+    let mPerSecond = 1000<m / second> // foo
 
     [<Measure>]
     type ProcessId =